v1.0.17.5 (Produção)
Alteração
- Revertendo alterações feitas na versão 1.0.17.4, pois estava ocorrendo Nullpointer além do fato do J2 ter continuado sem conseguir ler o qr code.
