+++ app/build.gradle
@@ -7,8 +7,8 @@ android {
         applicationId "br.com.home.maildeliveryjfsteel"
         minSdkVersion 14
         targetSdkVersion 25
-        versionCode 35
-        versionName "1.0.17.4"
+        versionCode 36
+        versionName "1.0.17.5"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
     }
